requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
faster-whisper>=1.0.0
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
